+++ misk/misk/web/tabs/example/src/components/TabComponent.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import styled from "react-emotion"
+import styled from "styled-components"
 
 export interface IComponentProps {
   data: IData